Revert to the state before all paroli changes
Some explanation : When we started the paroli project, we decided to work in
tichy repository, without branching (that is modify tichy to slowly turn it into
paroli) So we started to commit changes into tichy, that made it into a unstable
state.

The problem is that at some point we created a new repository for paroli, where
we imported all the files from the current branch of tichy. After this we
stopped working in tichy repository, letting it be in a unstabble state and
unmaintained.

This commit reverts tichy into the state it was prior to paroli integration. So
that tichy is freezed in this current state and we don't touch it anymore.

People should use paroli repository instead of tichy.